+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -403,7 +403,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KQ70ETS860YYW972QPWM2AHJ"),
+  _id: z.string().default("01KQ7Q5D4HNF05TW91GQJF5XMH"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,