+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-27T08:21:11.069Z",
+    "2026-04-27T14:57:59.356Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {