+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -176,7 +176,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KQ70ETS6CJZ82F4C0QV6YH6D"),
+  _id: z.string().default("01KQ7Q5D4F5DV21S7RW2PVNH03"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,