+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -152,7 +152,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KQ70ETSFCQ5CKSSZS16JJJNZ"),
+  _id: z.string().default("01KQ7Q5D4KDP5FEWY09K4900PZ"),
   display_name: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$inboundSchema,